--- src/actions/authActionTypes.js
@@ -1,2 +1,0 @@
-export const INJECT = 'INJECT_USER';
-export const LOGOUT = 'LOGOUT_USER';
