+++ src/actions/types.js
@@ -1,5 +1,8 @@
-export const PULLS_REQUEST = 'PULLS_REQUEST'
-export const PULLS_SUCCESS = 'PULLS_SUCCESS'
-export const PULLS_FAILURE = 'PULLS_FAILURE'
+export const PULLS_REQUEST = 'PULLS_REQUEST';
+export const PULLS_SUCCESS = 'PULLS_SUCCESS';
+export const PULLS_FAILURE = 'PULLS_FAILURE';
 
-export const GET_REPOSITORIES = 'get_repositories'
+export const GET_REPOSITORIES = 'get_repositories';
+
+export const INJECT = 'INJECT_USER';
+export const LOGOUT = 'LOGOUT_USER';
