+++ src/components/authentication/signin.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+class SignIn extends Component {
+ 
+  render() {
+    return (
+      <div>
+        PR Dashboard
+      </div>
+    )
+  }
+}
+
+export default SignIn
