+++ src/components/layout/header/component_dashboard_header.js
@@ -14,6 +14,8 @@ class DashboardHeader extends Component {
 	handleLogIn = () => {
 		const msg = loginTab(`https://pr-dashboard-server.herokuapp.com/v1/auth/github`)
 		msg.then(user => {
+			console.log('USER', user);
+			
 			this.props.authActions.injectUser(user)
 		})
 	}
@@ -44,6 +46,8 @@ class DashboardHeader extends Component {
 						</button>
 					</form>
 				</div>
+				
+				<a href={`${config.serverUrl}/auth/github`} target="_blank">link</a>
 
 				<div className="header__config">
 					<button className="header__config-button" onClick={this.handleLogIn}>
@@ -57,10 +61,10 @@ class DashboardHeader extends Component {
 	}
 }
 
-// const mapStateToProps = authentication => ({
-// 	isAuthenticated: authentication[USER_STATE_KEY].isAuthenticated,
-// 	currentUser: authentication[USER_STATE_KEY].currentUser
-// })
+const mapStateToProps = authentication => ({
+	isAuthenticated: authentication[USER_STATE_KEY].isAuthenticated,
+	currentUser: authentication[USER_STATE_KEY].currentUser
+})
 
 const mapDispatchToProps = dispatch => ({
 	authActions: bindActionCreators(authActions, dispatch)
